@@ -4,31 +4,27 @@
 	"private": true,
 	"dependencies": {
 		"@americanexpress/react-seo": "^1.4.2",
-		"@emotion/react": "^11.4.1",
-		"@emotion/styled": "^11.3.0",
+		"@emailjs/browser": "^3.11.0",
+		"@emotion/react": "^11.11.1",
+		"@emotion/styled": "^11.11.0",
 		"@mui/material": "^5.0.3",
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^11.2.7",
-		"@testing-library/user-event": "^12.8.3",
-		"emailjs-com": "^3.2.0",
-		"framer-motion": "^4.1.17",
+		"@tsparticles/react": "^3.0.0",
+		"@tsparticles/slim": "^3.0.2",
+		"framer-motion": "^10.16.16",
 		"ga-4-react": "^0.1.281",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-facebook-pixel": "^1.0.4",
-		"react-icons": "^4.3.1",
-		"react-router-dom": "^5.3.0",
-		"react-scripts": "4.0.3",
-		"react-tsparticles": "^1.35.4",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-icons": "^4.12.0",
+		"react-router-dom": "^6.21.1",
+		"react-scripts": "^5.0.1",
 		"semantic-ui-css": "^2.4.1",
 		"semantic-ui-react": "^2.0.4",
-		"typewriter-effect": "^2.18.2",
-		"web-vitals": "^1.1.2"
+		"typewriter-effect": "^2.21.0"
 	},
 	"devDependencies": {
-		"immer": "^9.0.6",
-		"postcss": "^8.4.5",
-		"node-forge": "^1.2.1"
+		"immer": "^10.0.3",
+		"node-forge": "^1.3.1",
+		"postcss": "^8.4.32"
 	},
 	"scripts": {
 		"start": "react-scripts start",
