@@ -1,20 +1,20 @@
 import { useEffect } from 'react';
-import { useHistory } from 'react-router';
+import { useNavigate } from 'react-router-dom';
 import { motion } from 'framer-motion';
 import { AiOutlineCloseCircle } from 'react-icons/ai';
 import { Button, Divider, Icon, Image } from 'semantic-ui-react';
 import SeoSettings from '../../seo/SeoSettings';
-import ReactPixel from 'react-facebook-pixel';
+// import ReactPixel from 'react-facebook-pixel';
 
-ReactPixel.pageView();
+// ReactPixel.pageView();
 
 const ProjectCard = ({ project }) => {
-	const history = useHistory();
+	const navigate = useNavigate();
 
 	const projectUrl = project.projectName.replaceAll(' ', '-').toLowerCase();
 
 	const handleClose = () => {
-		history.push('/portfolio');
+		navigate('/portfolio');
 	};
 
 	useEffect(() => {
