@@ -4,15 +4,15 @@ import { motion } from 'framer-motion';
 import { AiOutlineCloseCircle } from 'react-icons/ai';
 import { Icon, Image } from 'semantic-ui-react';
 import SeoSettings from '../../seo/SeoSettings';
-import ReactPixel from 'react-facebook-pixel';
+// import ReactPixel from 'react-facebook-pixel';
 
 const pixelOptions = {
 	autoConfig: true,
 	debug: false,
 };
 
-ReactPixel.init('258173202988755', pixelOptions);
-ReactPixel.pageView();
+// ReactPixel.init('258173202988755', pixelOptions);
+// ReactPixel.pageView();
 
 const LambdaCert = () => {
 	const history = useHistory();
